Add CurlBodyProps component and update ExecutePanel
