+++ components/layout/execute/ExecutePanel.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import React, { memo, useEffect, useState } from "react";
-import { CurlCommand, ExecuteHeader, ExecuteResponseExample } from "@/components";
+import { CurlBodyProps, CurlCommand, ExecuteHeader, ExecuteResponseExample } from "@/components";
 import { NotoSans } from "@/lib/assets";
 import { Controller, Endpoint, Project } from "@/lib/types";
 import { processQueryParameters, processRequestBody, processUrlParameters } from "@/lib/utils/generateCurlCommand";
@@ -14,23 +14,24 @@ interface Props {
 }
 
 function Component({ projectData, controllerData, selected, setSelected }: Props) {
-  const [curlCommand, setCurlCommand] = useState("");
+  const [startCommand, setStartCurlCommand] = useState("");
+  const [headers, setHeaders] = useState("");
   const [bodyProps, setBodyProps] = useState<Record<string, string>>();
+  const [paramsProps, setParamsProps] = useState<Record<string, string>>();
   const styles = selected ? "flex-1" : "flex-0";
 
   useEffect(() => {
     if (!selected) return;
-    const url = `${projectData.serverUrl}${controllerData.basePath ? "/" + controllerData.basePath : ""}${
-      selected.path
-    }`;
+    const url = `${projectData.serverUrl}${controllerData.basePath ? "/" + controllerData.basePath : ""}`;
     const { body, params, query } = selected.request;
     const { example } = selected.response;
 
     let processedUrl = url;
-    let curlCommand = `curl -X ${selected.method} ${processedUrl} \\\n`;
+    let startCommand = `curl -X ${selected.method} ${processedUrl}`;
 
     if (params) {
-      processedUrl = processUrlParameters(processedUrl, params.properties, example);
+      const requestParams = processUrlParameters(params.properties, example);
+      setParamsProps(requestParams);
     }
 
     if (query) {
@@ -40,13 +41,15 @@ function Component({ projectData, controllerData, selected, setSelected }: Props
     if (body) {
       const requestBody = processRequestBody(body.properties, example);
       setBodyProps(requestBody);
-      curlCommand += `-H "Content-Type: application/json" \\\n`;
+      setHeaders((prev) => prev + `-H "Content-Type: application/json" \\\n`);
     }
 
-    setCurlCommand(curlCommand);
+    setStartCurlCommand(startCommand);
 
     return () => {
       setBodyProps(undefined);
+      setParamsProps(undefined);
+      setHeaders("");
     };
   }, [selected]);
 
@@ -66,27 +69,12 @@ function Component({ projectData, controllerData, selected, setSelected }: Props
               <span>{selected.method}</span>
               <span>/{controllerData.basePath + selected.path}</span>
             </div>
-            <div className="flex flex-col gap-4">
-              {bodyProps &&
-                Object.entries(bodyProps).map(([key, value]) => {
-                  return (
-                    <div key={key} className="flex flex-col gap-2">
-                      <div className="text-2 text-white">{key}</div>
-                      <input
-                        onChange={(e) => {
-                          const newProps = { ...bodyProps };
-                          newProps[key] = e.currentTarget.value;
-                          setBodyProps(newProps);
-                        }}
-                        className="w-full py-4 px-8 rounded-sm outline-none border-none bg-gray-800 text-1 text-white"
-                        placeholder={value}
-                        type="text"
-                      />
-                    </div>
-                  );
-                })}
-            </div>
-            <CurlCommand command={curlCommand} formattedBody={JSON.stringify(bodyProps, null, 2) || undefined} />
+            <CurlBodyProps bodyProps={bodyProps} setBodyProps={setBodyProps} />
+            <CurlCommand
+              startCommand={startCommand}
+              headers={headers}
+              formattedBody={JSON.stringify(bodyProps, null, 2) || undefined}
+            />
             <ExecuteResponseExample endpoint={selected} />
           </div>
         </>
