+++ components/layout/execute/index.ts
@@ -1,4 +1,5 @@
+export * from "./CurlBodyProps";
+export * from "./CurlCommand";
 export * from "./ExecutePanel";
 export * from "./ExecuteResponseExample";
-export * from "./CurlCommand";
 export * from "./ExecuteHeader";
